feat: login UI merge 및 수정 & layout 분리
