+++ src/app/(auth)/layout.module.css
@@ -1,0 +1,20 @@
+.root {
+  height: 100vh;
+}
+
+.header {
+  height: 180px;
+
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.header > * {
+  margin: auto;
+}
+
+.logoContainer {
+  width: 18rem;
+  height: auto;
+}
