+++ src/app/(auth)/login/login.module.css
@@ -1,0 +1,132 @@
+.Wrapper > * {
+  margin: auto;
+  width: 28.75rem;
+}
+
+.inputContainer {
+  padding: 1.25rem 1.75rem;
+
+  border: 1px solid #c6c6c6;
+  border-radius: 6px;
+
+  box-shadow: 0 5px 8px 0 rgba(68, 68, 68, 0.08);
+}
+
+.inputWrapper {
+  border: 1px solid #dadada;
+  border-radius: 6px;
+
+  box-shadow: 0 2px 6px 0 rgba(68, 68, 68, 0.08);
+}
+
+.inputLine {
+  padding: 0.75rem 1.125rem;
+}
+
+.inputLine > * {
+  vertical-align: middle;
+}
+
+.divisionLine {
+  height: 1px;
+  background-color: #dadada;
+}
+
+.idWrapper {
+  border-radius: 6px 6px 0 0;
+}
+
+.pwWrapper {
+  border-radius: 0 0 6px 6px;
+}
+
+.input {
+  font-size: 1rem;
+  margin-left: 8px;
+  width: 90%;
+}
+
+.keep_check {
+  margin-top: 0.875rem;
+}
+
+.keep_check > p {
+  font-size: 14px;
+  color: #777777;
+
+  margin-left: 8px;
+}
+
+.button {
+  width: 100%;
+
+  padding: 0.875rem 0;
+
+  border-radius: 6px;
+
+  background-color: #003b71;
+
+  font-size: 1.25rem;
+  font-weight: bold;
+  color: white;
+
+  margin-top: 2.375rem;
+}
+
+.bottomBar {
+  display: flex;
+  justify-content: center;
+
+  margin-top: 1.25rem;
+}
+
+.navWrapper * {
+  /* display: flex;
+  align-items: center; */
+
+  color: #888888;
+  font-size: 0.875rem;
+
+  vertical-align: middle;
+}
+
+.social {
+  margin-top: 3.75rem;
+
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+
+  height: auto;
+}
+
+.socialText {
+  /* display: flex;
+  justify-content: center;
+  align-items: center; */
+
+  width: 100%;
+  height: auto;
+
+  font-size: 12px;
+  color: #888888;
+
+  margin-bottom: 30px;
+}
+
+.socialText > * {
+  display: inline-block;
+  vertical-align: middle;
+}
+
+.socialText > p {
+  width: 40%;
+  text-align: center;
+}
+
+.divideLine {
+  width: 30%;
+  height: 0.1px;
+
+  background-color: #888888;
+}
