--- src/app/login/login.module.css
@@ -1,194 +1,0 @@
-.Wrapper {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    height: 100vh;
-}
-
-.header {
-    width: 1500px;
-    height: 180px;
-    
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-
-.logoContainer {
-    width: 18rem;
-    height: auto;
-}
-
-.loginWrapper {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-
-    width: 1500px;
-    height: auto;
-}
-
-.loginWrapper > :first-child {
-    position: relative;
-    display: flex;
-    flex-direction: column;
-
-    width: 460px;
-    height: 255px;
-
-    border: 1px solid #C6C6C6;
-    border-radius: 6px;
-
-    padding-top: 20px;
-    padding-bottom: 20px;
-    padding-left: 29px;
-    padding-right: 29px;
-}
-
-.id_pw_Wrapper {
-    width: 402px;
-    height: 94px;
-}
-
-.idWrapper {
-    width: 402px;
-    height: 47px;
-
-    display: flex;
-    align-items: center;
-
-    border: 1px solid #DADADA;
-    border-radius: 6px 6px 0 0;
-
-    padding-left: 20px;
-}
-
-.pwWrapper {
-    width: 402px;
-    height: 47px;
-
-    display: flex;
-    align-items: center;
-
-    border: 1px solid #DADADA;
-    border-radius: 0 0 6px 6px;
-
-    padding-left: 20px;
-}
-
-.input {
-    width: 80%;
-
-    font-size: 16px;
-
-    padding-left: 10px;
-}
-
-.keep_check {
-    position: relative;
-    display: flex;
-    flex-direction: row;
-
-    height: 14px;
-
-    margin-top: 10px;
-}
-
-.keep_check > p {
-    font-size: 14px;
-    color: #777777;
-
-    margin-left: 8px;
-}
-
-.button {
-    position: absolute;
-    bottom: 20px;
-
-    width: 402px;
-    height: 50px;
-
-    border-radius: 6px;
-
-    background-color: #003B71;
-
-    font-size: 20px;
-    font-weight: bold;
-    color: white;
-
-    margin-top: 20px;
-}
-
-.bottomBar {
-    display: flex;
-    flex-direction: row;
-
-    margin-top: 25px;
-    margin-bottom: 60px;
-}
-  
-.navWrapper {
-    display: flex;
-    align-items: center;
-
-    color: #888888;
-    font-size: 14px;
-}
-
-.social {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-
-    width: 460px;
-    height: auto;
-}
-
-.socialText {
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    align-items: center;
-
-    width: 400px;
-    height: auto;
-
-    font-size: 12px;
-    color: #888888;
-
-    margin-bottom: 30px;
-}
-
-.divideLine {
-    width: 100px;
-    height: 0.1px;
-
-    background-color: #888888;
-
-    margin: 0 10px;
-}
-
-.kakaoContainer {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-
-    width: 340px;
-    height: 44px;
-
-    border-radius: 12px;
-
-    background-color: #FAE500;
-
-    font-size: 14px;
-    font-weight: bold;
-    color: #000000 85%;
-
-    margin-bottom: 40px;
-}
-
-.kakaoContainer img {
-    margin-right: 10px;
-}
